+++ src/components/AnimateSharedLayout/utils/batcher.ts
@@ -1,4 +1,4 @@
-import { flushSync } from "framesync"
+import sync, { flushSync } from "framesync"
 import { VisualElement } from "../../../render/types"
 import { compareByDepth } from "../../../render/utils/compare-by-depth"
 import { Presence, SyncLayoutBatcher, SyncLayoutLifecycles } from "../types"
@@ -62,7 +62,21 @@ export function createBatcher(): SyncLayoutBatcher {
             flushSync.preRender()
             flushSync.render()
 
+            /**
+             * Schedule a callback at the end of the following frame to assign the latest projection
+             * box to the prevViewportBox snapshot. Once global batching is in place this could be run
+             * synchronously. But for now it ensures that if any nested `AnimateSharedLayout` top-level
+             * child attempts to calculate its previous relative position against a prevViewportBox
+             * it will be against its latest projection box instead, as the snapshot is useless beyond this
+             * render.
+             */
+            sync.postRender(() => order.forEach(assignProjectionToSnapshot))
+
             queue.clear()
         },
     }
+}
+
+function assignProjectionToSnapshot(child: VisualElement) {
+    child.prevViewportBox = child.projection.target
 }