+++ CHANGELOG.md
@@ -6,7 +6,7 @@ Framer Motion adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 
--   Being more conservative with when we perform relative projection.
+-   Layout projection improvements.
 
 ## [4.1.4] 2021-04-15